line eater for CS

+++ src/d_ivl_ports.h
@@ -175,7 +175,7 @@ public:
   static int count()			{return _count;}
 //  vpiHandle H; // ?
   LOGICVAL lvfromivl;
-  double edge(int, unsigned delay=0);
+  unsigned edge(int, unsigned delay=0);
   void	   qe(); //{tr_queue_eval();}
 private:
   bool	   tr_eval_digital();
